requirements.txt
fastapi
uvicorn
opencv-python
pillow
python-multipart
numpy
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
bcrypt
